IGNITE-16520 Refactor IgniteCliInterfaceTest
