+++ modules/cli/src/integrationTest/java/org/apache/ignite/internal/cli/commands/connect/ItConnectCommandTest.java
@@ -106,7 +106,6 @@ class ItConnectCommandTest extends ItConnectToClusterTestBase {
         assertThat(getPrompt()).isEqualTo("[" + nodeName() + "]> ");
 
         // When connect again
-        resetOutput();
         execute("connect");
 
         // Then
