+++ modules/cli/src/test/java/org/apache/ignite/internal/cli/commands/IgniteCliInterfaceTestBase.java
@@ -1,0 +1,50 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.ignite.internal.cli.commands;
+
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockserver.integration.ClientAndServer;
+import org.mockserver.junit.jupiter.MockServerExtension;
+
+/**
+ * Base class for testing CLI interface.
+ */
+@ExtendWith(MockServerExtension.class)
+public class IgniteCliInterfaceTestBase extends CliCommandTestBase {
+    protected static ClientAndServer clientAndServer;
+
+    protected static String mockUrl;
+
+    @Override
+    protected Class<?> getCommandClass() {
+        return TopLevelCliCommand.class;
+    }
+
+    @BeforeAll
+    static void initMockServer(ClientAndServer clientAndServer) {
+        IgniteCliInterfaceTestBase.clientAndServer = clientAndServer;
+        mockUrl = "http://localhost:" + clientAndServer.getPort();
+    }
+
+    @BeforeEach
+    void resetMockServer() {
+        clientAndServer.reset();
+    }
+}
